Clean agenda restrict logic in org-roam-ql-view--agenda
This commit refactors org-roam-ql-view--agenda by removing legacy
handling of org-agenda-restrict and related buffer narrowing
logic. The agenda view now provides a message if no nodes are
found. This simplifies the code and removes unused restriction logic
that was previously copied from org-ql-search-block.

When I figure out how to handle the restrict logic, this can be
revisited.