Update documentation on TYPE and COMBINE options for links
Update documentation on TYPE and COMBINE keyword args for link
expansions. org-roam-ql--expand-forwardlinks and
org-roam-ql--expand-backlinks now default COMBINE to :or and forward/
back links docstrings were rewritten for clarity. README examples for
backlink-to and backlink-from were updated to show the new signatures
(&key (TYPE "id") (COMBINE :or)).

May break anything that depends on the default COMBINE being set to
:and